% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{biomarker_score}
\alias{biomarker_score}
\title{Compute the gamma biomarker score for the threshold. Also uses gamma validation.}
\usage{
biomarker_score(biomarkers, gamma)
}
\arguments{
\item{biomarkers}{A data frame (or matrix) of biomarker values.}

\item{gamma}{A vector of biomarker coefficients.}
}
\value{
A vector containing \code{1 + biomarkers \%*\% gamma}.
}
\description{
Compute the gamma biomarker score for the threshold. Also uses gamma validation.
}
