% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaned_data}
\alias{cleaned_data}
\title{Real and Anonymus cleaned patient data contatining all information necessary to use the Single-Index-Model}
\format{
\subsection{\code{cleaned_data}}{

A data frame with 569 rows and 53 columns:
\describe{
\item{time}{Failure times}
\item{status}{A binary 1 for death and 0 for censoring}
\item{trt}{A binary 1 for "recieved treatment" and 0 for "did not recieve treatment"}
\item{bio1, bio2, ...}{Biomarkers}
}
}
}
\source{
Private
}
\usage{
cleaned_data
}
\description{
Real and Anonymus cleaned patient data contatining all information necessary to use the Single-Index-Model
}
\keyword{datasets}
