% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siutils.R
\name{cmb_biomarker}
\alias{cmb_biomarker}
\title{Combine biomarkers with gamma linearly based on the single-index model}
\usage{
cmb_biomarker(biomarkers, gamma)
}
\arguments{
\item{biomarkers}{A data frame of row(i)=subjects_i, col(j)=biomarker_j.}

\item{gamma}{A vector of biomarker effects on individuals sensitvity}
}
\value{
Combined threshold value
}
\description{
Combine biomarkers with gamma linearly based on the single-index model
}
