% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{cox_threshold_loglik}
\alias{cox_threshold_loglik}
\title{Evaluates the Cox partial log likelihood.}
\usage{
cox_threshold_loglik(
  x,
  y,
  beta,
  gamma,
  treatment_col = 1,
  biomarker_cols = NULL
)
}
\arguments{
\item{x}{A data frame of covariates with \code{treatment} column and one or more biomarker (|treatment|bio1|...|bion|)
columns.}

\item{y}{A \code{survival::Surv} response object sorted in descending time order.}

\item{beta}{A vector of regression coefficients for treatment,
biomarker subgroup, and their interaction term.}

\item{gamma}{A vector of biomarker coefficients.}

\item{treatment_col}{Column name or index for the treatment indicator.}

\item{biomarker_cols}{Optional column names or indices for biomarkers.}
}
\value{
A scalar log partial likelihood.
}
\description{
Evaluates the Cox partial log likelihood.
}
