% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{draw_gamma_uniform}
\alias{draw_gamma_uniform}
\title{Draw gamma from a uniform prior distribution.}
\usage{
draw_gamma_uniform(n_biomarkers, min = -1, max = 1)
}
\arguments{
\item{n_biomarkers}{Number of biomarker coefficients (=gamma components) to draw.}

\item{min}{Lower bound vector (or scalar if identical) for the uniform prior.}

\item{max}{Upper bound vector (or scalar if identical) for the uniform prior.}
}
\value{
A numeric gamma vector.
}
\description{
Draw gamma from a uniform prior distribution.
}
