% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBM_col}
\alias{getBM_col}
\title{Basic helper to extract Biomarker index}
\usage{
getBM_col(x)
}
\arguments{
\item{x}{A data frame of study data (see format elsewhere)}
}
\value{
Index where biomarkers are foubd
}
\description{
Basic helper to extract Biomarker index
}
