% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{initialize_beta}
\alias{initialize_beta}
\title{Initialize beta with the Cox fit.}
\usage{
initialize_beta(x, y, gamma, treatment_col = 1, biomarker_cols = NULL)
}
\arguments{
\item{x}{A data frame of covariates with \code{treatment} column and one or more biomarker (|treatment|bio1|...|bion|)
columns.}

\item{y}{A \code{survival::Surv} response object.}

\item{gamma}{A vector of biomarker coefficients.}

\item{treatment_col}{Column name or index for the treatment indicator.}

\item{biomarker_cols}{Optional column names or indices for biomarkers.}
}
\value{
beta coefficients from the Cox model.
}
\description{
Initialize beta with the Cox fit.
}
