% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{match_columns}
\alias{match_columns}
\title{Match requested columns to numeric indices.}
\usage{
match_columns(x, cols)
}
\arguments{
\item{x}{A data frame or matrix.}

\item{cols}{Column names or indices.}
}
\value{
Integer column indices.
}
\description{
Match requested columns to numeric indices.
}
\keyword{internal}
