% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{propose_gamma}
\alias{propose_gamma}
\title{Propose an updated gamma vector.}
\usage{
propose_gamma(gamma, proposal_sd = 0.05)
}
\arguments{
\item{gamma}{Current gamma vector.}

\item{proposal_sd}{Random-walk proposal standard deviation.}
}
\value{
A proposed gamma vector.
}
\description{
Propose an updated gamma vector.
}
