% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{repeat_parameter}
\alias{repeat_parameter}
\title{Repeat a scalar distribution parameter.}
\usage{
repeat_parameter(value, n)
}
\arguments{
\item{value}{Scalar or vector distribution parameter.}

\item{n}{Required parameter length.}
}
\value{
A vector of length \code{n}.
}
\description{
Repeat a scalar distribution parameter.
}
\keyword{internal}
