% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{simu_data}
\alias{simu_data}
\title{Simulate study data based on predetermined coefficients (beta, gamma)}
\usage{
simu_data(n, beta, gamma, h0 = 1)
}
\arguments{
\item{n}{The number of observations (subjects).}

\item{beta}{A vector of effects: 1-Trt & not sensitive. 2-Sensitive & no trt. 1+2+3-Sensitive & trt}

\item{gamma}{A vector of biomarker effects on individuals sensitvity}
}
\value{
A data frame of time | status | trt | sensitivity | biomarker 1 | ... | biomarker n |
}
\description{
Simulate study data based on predetermined coefficients (beta, gamma)
}
