% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{summarize_mcmc}
\alias{summarize_mcmc}
\title{Summarize posterior samples.}
\usage{
summarize_mcmc(posterior, alpha = 0.05, method = "mean")
}
\arguments{
\item{posterior}{A posterior sample list from \code{fit_threshold_mcmc()} (or \verb{from fit_threshold_model()}).}

\item{alpha}{Error probability for equal-tailed credible intervals.}

\item{method}{Point estimate method, either \code{"mean"} or \code{"median"}.}
}
\value{
A list of posterior point estimates and credible intervals.
}
\description{
Summarize posterior samples.
}
