% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data.R
\name{summary_data}
\alias{summary_data}
\title{Summarize data}
\usage{
summary_data(data)
}
\arguments{
\item{data}{A data frame: time | status | trt | sensitivity | biomarker 1 | ... | biomarker n |.}
}
\value{
A list:
}
\description{
Summarize data
}
