% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siutils.R
\name{thresholding}
\alias{thresholding}
\title{Combine biomarkers with gamma linearly based on the single-index model}
\usage{
thresholding(x, gamma)
}
\arguments{
\item{x}{A data frame with trt and biomarker values.}

\item{gamma}{A vector of biomarker effects on individuals sensitvity}
}
\value{
A data frame: trt | sensitivity | product of trt*sensitivity
}
\description{
Combine biomarkers with gamma linearly based on the single-index model
}
