% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{validate_gamma}
\alias{validate_gamma}
\title{Validate gamma dimensions for calculations and then return it if it passes.}
\usage{
validate_gamma(gamma, n_biomarkers)
}
\arguments{
\item{gamma}{A vector of biomarker coefficients.}

\item{n_biomarkers}{Integer number of biomarkers.}
}
\value{
\code{gamma} as a numeric vector.
}
\description{
Validate gamma dimensions for calculations and then return it if it passes.
}
