% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{default_mcmc_control}
\alias{default_mcmc_control}
\title{Default MCMC parameters to control sampling and gamma's distribution.}
\usage{
default_mcmc_control(
  samples = 5000,
  burn_in = 3000,
  thin = 10,
  gamma_mean = 0,
  gamma_sd = 0.1,
  gamma_proposal_sd = 0.05,
  gamma_prior = c("normal", "uniform"),
  gamma_min = -1,
  gamma_max = 1
)
}
\arguments{
\item{samples}{Number of posterior draws.}

\item{burn_in}{Number of burn-in (throw away) iterations.}

\item{thin}{Keep every \code{thin}-th draw (after \verb{burn-in}).}

\item{gamma_mean}{Mean vector or scalar for the normal gamma prior.}

\item{gamma_sd}{Standard deviation vector or scalar for the normal gamma
prior.}

\item{gamma_proposal_sd}{Random-walk proposal standard deviation for gamma.}

\item{gamma_prior}{Gamma prior distribution, either \code{"normal"} or
\code{"uniform"}.}

\item{gamma_min}{Lower bound vector or scalar for the uniform gamma prior.}

\item{gamma_max}{Upper bound vector or scalar for the uniform gamma prior.}
}
\value{
A named list of MCMC controls.
}
\description{
Default MCMC parameters to control sampling and gamma's distribution.
}
