% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{draw_gamma}
\alias{draw_gamma}
\title{Draw gamma from prior distribution.}
\usage{
draw_gamma(n_biomarkers, mean = 0, sd = 0.1)
}
\arguments{
\item{n_biomarkers}{Number of biomarker coefficients (=gamma components) to draw.}

\item{mean}{Mean vector (or scalar if identical) for the normal prior.}

\item{sd}{Standard deviation vector (or scalar if identical) for the normal prior.}
}
\value{
A numeric gamma vector.
}
\description{
Draw gamma from prior distribution.
}
