% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_index_utils.R
\name{log_gamma_prior}
\alias{log_gamma_prior}
\title{Evaluate the gamma prior density.}
\usage{
log_gamma_prior(gamma, mean = 0, sd = 0.1)
}
\arguments{
\item{gamma}{A vector of biomarker coefficients.}

\item{mean}{Mean vector (or scalar if identical) for the normal prior.}

\item{sd}{Standard deviation vector (or scalar if identical) for the normal prior.}
}
\value{
The log prior density.
}
\description{
Evaluate the gamma prior density.
}
