% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_step}
\alias{mcmc_step}
\title{Metropolis-Hastings update.}
\usage{
mcmc_step(
  x,
  y,
  beta,
  gamma,
  control,
  lambda = 0,
  treatment_col = 1,
  biomarker_cols = NULL
)
}
\arguments{
\item{x}{A data frame of covariates with \code{treatment} column and one or more biomarker (|treatment|bio1|...|bion|)
columns.}

\item{y}{A \code{survival::Surv} response object sorted in descending time order.}

\item{beta}{A 3-dimensional vector of the current regression coefficients.}

\item{gamma}{An n-dimensional vector of the current biomarker coefficients.}

\item{control}{A list from \code{default_mcmc_control()}.}

\item{lambda}{Nonnegative lasso penalty for gamma.}

\item{treatment_col}{Column name or index for the treatment indicator.}

\item{biomarker_cols}{Optional column names or indices for biomarkers.}
}
\value{
A list containing updated \code{beta}, \code{gamma}, and \code{log_likelihood}.
}
\description{
Metropolis-Hastings update.
}
